server: improve port binding, --port-range option
Previously, we used a fixed port and guessed that it'll be ready ~200ms
after the server start.

I think I figured out how to bind to a port before starting a server.
Now, we can handle a port range and we can start a browser after we know
the port is valid.