webapp: fix CSS bug with large files in pinned mode
I do not fully understand the bug. IIRC, the previous code worked fine
for a while, but a few months ago the scroll bars stopped working. The
fix is mostly AI generated, which took several tries.

The proximate cause is that the setting the `height` value in the diff
of this comment to a *percentage* value (or `calc` involving such a
value) now makes the height very large, even though the parent
`<details>` element has a reasonable height.

I think this is some new CSS weirdness with `<details>`, as happened
before. It seems cursed, a custom `<div>` with custom JS would probably
be better. Or, perhaps, the problem is my workarounds for the old
`<details>` weirdness; this CSS was never sane, and perhaps now a
saner solution is available. TODO: try to make it at least a little
saner.